changed to an imgui implementation
Importantly, I needed to include the headers and cpp source code for the imgui impl glfw and vulkan, because I don't think they're naturally in conancenter. Otherwise, this process was very straightforward. I just added the relevant things to the conanfile and did the same in the meson.build and it all came together mostly. This doesn't work on apple silicon yet because of awesome portability things I haven't remembered how to deal with, but other than that this is really close to working.